refactor(blue): [unit] split video_creator
