+++ src/tdd_in_python/outside_in_04/__init__.py
@@ -1,5 +1,5 @@
 from .app import app as app
 from .database import Base as Base
 from .database import engine as engine
-from .video_creator import Video as Video
+from .video import Video as Video
 from .video_creator import VideoCreator as VideoCreator
