+++ src/tdd_in_python/outside_in_04/video.py
@@ -1,0 +1,11 @@
+from sqlalchemy.orm import Mapped
+from sqlalchemy.orm import mapped_column
+
+from .database import Base
+
+
+class Video(Base):
+    __tablename__ = "videos"
+
+    id: Mapped[int] = mapped_column(primary_key=True)
+    title: Mapped[str]
