+++ src/tdd_in_python/outside_in_04/video_creator.py
@@ -1,33 +1,7 @@
-import abc
 import re
 
-from sqlalchemy.orm import Mapped
-from sqlalchemy.orm import mapped_column
-from sqlalchemy.orm import Session
-
-from .database import Base
-
-
-class Video(Base):
-    __tablename__ = "videos"
-
-    id: Mapped[int] = mapped_column(primary_key=True)
-    title: Mapped[str]
-
-
-class VideoRepository(abc.ABC):
-    @abc.abstractmethod
-    def save(self, video: Video):
-        raise NotImplementedError
-
-
-class SQLAlchemyVideoRepository(VideoRepository):
-    def __init__(self, session: Session):
-        self.session = session
-
-    def save(self, video: Video):
-        self.session.add(video)
-        self.session.commit()
+from .video import Video
+from .video_repository import VideoRepository
 
 
 class VideoCreator:
